Improve some of the island maps where better upstream versions exist
For Polynesia, the labels and circles were removed as they were barely
readable and slightly confusing.

The number of colours (using `convert -colors 255`) was reduced for
the Azores, Ceuta and Melilla without any discernable drop in quality.
Testing showed that reducing the number of colours for Polynesia would
cause a drop in quality, presumably because its map isn't "flat"
like the others.

(Terminology-wise I think that it's not strictly speaking
colour-depth, as the images are still 8-bit.  I suspect that the
significant drop in size is due to transitions at boundaries, where
there would normally be a large range of different colours, but most
of which aren't really necessary for the depiction of a smooth
transition.)

(I have no idea why using 255 colours results in a non-negligible gain
compared to 256 colours — AFAIK `-colors n` *doesn't* mean `n + 1`
colours.)