Add sources for the remaining maps
I'm leaving the statement in `LICENSE.md`, about the maps being
sourced from Wikimedia Commons, as the author of the majority of the
maps (TUBS) has asked:

"I'd greatly appreciate, that you attribute this media file to
Wikimedia Commons, if used outside Wikipedia or Commons."

For example, see here:

https://commons.wikimedia.org/wiki/File:Abkhazia_in_its_region.svg

I think that having the links to specific files in `sources.csv`, in
addition, is valuable for consistency and for our own organisational
convenience (e.g. checking for newer/better versions).