Add source for EU flag
According to:

https://commons.wikimedia.org/wiki/File:Flag_of_Europe.svg

only the implementation of the design, not the design itself, is in
the public domain. The copyright for the design lies with the Council
of Europe. They allow the use of the flag under the conditions
specified here:

https://www.coe.int/en/web/about-us/the-european-flag#column-5

I've summarised that as "Fair Use", which isn't quite right, as the
scope of the license is considerably broader than just "Fair Use".

Arguably, the same issue might hold for some of the other flags that
are recent enough not to be automatically in the public domain. The
Wikimedia editors addressed that explicitly for some (e.g. Montenegro
— according to Montenegrin laws the flag is public domain), but
seemingly not for others (e.g. Kosovo).