Use alternative source for English Channel
This results in better conformance with the IHO definition (and the
"common sense" one, as well).

https://doi.pangaea.de/10.1594/PANGAEA.777975

(This source wasn't used for most of the other sea maps, since:

a) Converting from KML to SHP (or geojson) was messy.

b) The limits of seas usually extended beyond the coast,
in-land (which was not visible in the SVG and had no effect on the
output PNG, but was still felt slightly "wrong").

c) Some of the seas had "incorrect" limits (e.g. Coral Sea, if I
remember correctly).)