Change the map of the Baltic Sea to follow the IHO 2002 draft
See:
http://wiki.geosys.ru/lib/exe/fetch.php/ru/portal/lib/iho/s23.los.ed4draft.2002.pdf#page=34

It now includes the Kattegat and the Danish Straits.

The current (old) map is just _wrong_ — it follows neither the IHO
1953 publication, nor the IHO 2002 draft, or any other logical limit
of the Baltic Sea.  It excludes the Kattegat and the Danish Straits,
like IHO 1953, but unlike IHO 2002, but also excludes some of the sea
between the Danish islands and Germany, which is part of the Baltic
Sea under even the most restrictive limits.

The reason for the old, incorrect map was because these were the best
boundaries that could be obtained, by combining various polygons, from
Natural Earth Data.  (Natural Earth Data has a polygon for "Internal
Denmark Waters", which combines some of the Danish straits and part of
the sea between the straits and Germany, while the polygon for the
Kattegat includes part of the Skagerrak.)

[Pangaea](https://doi.pangaea.de/10.1594/PANGAEA.777975) (under CC-BY)
is the data source for the new map.  I hadn't used it previously since
I was aiming to stay as close as possible to IHO 1953 rather than to
IHO 2002, but we've now explicitly switched to IHO 2002.  (Note that
the source claims to follow IHO 1953, but in the case of the Baltic
Sea it's not.)