Update CONTRIBUTING.md regarding editing guid.csv
I think that we've had issues with spreadsheets changing some of the
existing GUIDs to `Error:xxx` (in various languages) several times,
now, which is probably rather annoying for our contributors. (For the
latest see #538.)

Unfortunately, the CSV standard is under-defined and the "proper"
behaviour of cells starting with an equals sign isn't really
defined. (Python's CSV module thinks that a cell starting with an
equals sign is just a raw value; most spreadsheets try to interpret
such a cell as a formula.)