Add alternative name of Turkey
Fix #645.

This is a minor difference, but I think that it should be included
especially given the importance attached to this by the Turkish
government.

AFAICT Wikipedia in no other language includes Türkiye as a name in
the given language (even though they do often say "in Turkish: called
Türkiye"), so I think this should be in English only.