Rename East Timor to Timor-Leste
Fix #672.

Wikipedia changed the name in 2024-12:

https://en.wikipedia.org/w/index.php?title=Timor-Leste&oldid=1264890203

after a discussion and formal vote.

Change in English only (none of the other Wikipedias have changed the
article name).  (However, also change the parenthetical name in
Chinese.)

Note that for the Wikimedia sources of the map/flag, the upstream name
hasn't been changed, so the links in sources.csv are unchanged.