Update flag of Hungary based on Wikimedia (#705)
Fix #704.

AFAIU this is not a change due to external (legal) changes, but based
on what Wikipedia decided is the correct choice of flag colours.  (The
new colours are the legally-mandated colours.)

The new, updated flag has been on Wikipedia since 4 January 2023, so
it's relatively unlikely to be a temporary "edit war" version.