Combine "changes that prevent upgrading" section with previous one
In the context of erroring out/losing progress, the section in its
previous form is not currently true (we even ~recommend
adding/removing fields when switching between the extended and
experimental decks!).  I'm not actually sure whether it ever was true
— at least when using CrowdAnki (CrowdAnki has had the scary "map
cards and fields" dialog for a very long time).  When importing with
pure Anki (APKG) it quite likely was true, but Anki's imports are now
rather capable (and besides we don't really focus on APKG imports).

Just in case, though, I've kept the strong warning about comprehensive
data loss.

(The loss of the new field was and is true, but it doesn't prevent
upgrading and doesn't lead to other loss of progress.)

Co-authored-by: Axel Bocciarelli <axelboc@gmail.com>