Update flag of Afghanistan
This version is (according to Wikimedia) the more commonly used one.
It's been unchanged on Wikimedia since 2024-09 and it's been the
"main" one there (with some edit warring) since 2024-09.

Note however that the precise design of the Shahada (the text) is not
officially mandated, so AFAIU our previous version wasn't wrong, as
such.