Update flag of Albania
Apparently the "new" colours (updated in 2019, with some minor tweaks
later) better reflect the law from 2002:

https://upload.wikimedia.org/wikipedia/commons/d/d6/Ligji_8926_22.07.2002.pdf#page=4

which does seem to say that the red is to be
CMYK:(0,100,100,0) (i.e. #f00).

The Albanian president's page

https://president.al/simbolet/

also seems to use #f00, so probably Wikimedia is correct.

(I'm hesitant, because the red is apparently described as blood red,
which is not necessarily #f00; it's possible that the law/the document
decided to use the simplest red in the technical definition, out of
convenience, (despite it not being the actual "standard" shade),
Wikimedia/Wikipedia decided to strictly follow the law and now
everyone is copying Wikimedia/Wikipedia (e.g. the president's website
has the flag at:
https://president.al/wp-content/uploads/2023/09/Flag_of_Albania.png
(i.e. given the filename it was likely just converted from
Wikipedia).)  OTOH if the colour specified is "wrong" then the
Albanian Parliament could just change it.)