Update flags based on changes on Wikimedia (A-C) (partial) (#723)
None of the changes are major.
    
Three are due to "actual" external changes:
    
Azerbaijan's change is due to a modified law.  Aruba's probably also
is.  Austria's is due to changed guidelines (there's no "official"
colours, but the Ministry of Defense switched recommendations from
Pantone 032 C to Pantone 186 C in 2018).
    
Most of the rest are to better match existing external sources.  Many
are due to questions of how best to convert pantone to RGB.


* Make Afghanistan flag source consistent

This makes scripting slightly simpler.